add retry on insert operations
